refactor: split state.c into policy (state.c) and persistence (state_store.c)
state.c mixed two jobs in one 735-line file: the RAM state + policy (the
synced struct accessors, the brightness auto/manual model, the OS
pin/auto/host/detection ladder, idle style, glyph script/size, boot flags,
and the dirty-flag bookkeeping) and the EEPROM persistence (everything that
knows the poly_eeconf_t offsets and calls eeconfig_*).

The persistence half moves verbatim into state_store.c: save_user_settings,
save_user_latin, stamp_keymap_layers_fmt, save_user_eeconf,
save_user_mru_if_dirty, load_user_mru, load_user_eeconf, and the four
tail-byte writers (save_user_os / _glyph_script / _glyph_size /
_boot_flags, previously static — now declared in state.h with a "only
save_all_dirty calls these" note, since a direct call would bypass the
dirty bookkeeping). The store reads the values it persists through the
existing public getters (get_local_state, get_user_brightness,
get_idle_style, pack_auto_brightness, get_global_latin_table,
pack_os_state, get_glyph_script, get_glyph_size) plus one new
get_boot_flags(); nothing flows the other way, so the two halves meet only
at the state.h seam. state.c keeps the dirty flags and the save_all_dirty
orchestration — flags stay with the values they guard.

Behaviour-identical: same write contents, same order, same offsets;
declarations stay in state.h so every consumer is unchanged. The monolithic
POLYKYBD_DOOM=yes flavour's .data/.bss/.heap are byte-identical (3484 B
free heap, unchanged); split72 + split42 build, all 9 unit suites green
(117 tests), cog regeneration byte-identical, lint job + cppcheck clean.
